rewrite hosvd
In order to improve performance the hosvd function has been rewritten
Now it uses an algorithm that performs better, because the compression
is done earlier.
The old functions are renamed to hosvd2. The new function may only be
used with an relative tolerance, not with prescribed ranks.

For convenience also a function for forming a full tensor from a
Tucker Tensor is included.