fix hosvd and introduce some functions for convenience
The hosvd was not working correct after the last change. Now a proper
check for the bounds of the truncation index/rank is introduced. This
prevents from too large ranks, which results in wrong indices,
i.e. indexing columns of U that don't exist.

Add functions for getting the right-hand side without having to create
the vector xi explicitly.