add adaptive cross approximation
Algorithm for aca has been added. This currently only works with full
arrays.
For use with this also a function computing the maximum of a vector
except for some indices had to be implemented.
For better use need a function that does not rely on functionality
only an array implements, but also for more abstract functions, that
behave like arrays.